requirements.txt
requests
matplotlib
pillow
pycrypto
django-jsonfield
django-json-widget